coverage wip
disable mypy pre-commit

PRs should check mypy, commits should just 'work' to avoid hassle when saving work and moving between workstations

+++ tests/util.py
@@ -1,12 +1,16 @@
+import contextlib
 import os
 import tempfile
 from typing import Optional
 
 
 class TempFileHandle:
-    def __init__(self, suffix: Optional[str] = None):
-        with tempfile.NamedTemporaryFile("x", delete=False, suffix=suffix) as h:
-            self._filename = h.name
+    def __init__(self, suffix: Optional[str] = None, *, create: bool = True):
+        if create:
+            with tempfile.NamedTemporaryFile("x", delete=False, suffix=suffix) as h:
+                self._filename = h.name
+        else:
+            self._filename = None
 
     @property
     def path(self):
@@ -20,6 +24,7 @@ class TempFileHandle:
 
     def __exit__(self, exc_type, exc_val, exc_tb):
         try:
-            os.unlink(self._filename)
+            if self._filename is not None:
+                os.unlink(self._filename)
         except Exception as e:
             print(e)
